api/requirements.txt
uvicorn
fastapi
psycopg2-binary
python-multipart
mutagen